Removing duplicate test code: checking resolved subgroup host identification done.
