+++ testing/check_tracing_particles_resolved_subgroups.py
@@ -1,191 +1,44 @@
 #!/bin/env python
-from tqdm import tqdm
+
+# Retrieve helper functions, without having to define an __init__.py 
+import sys
+sys.path.append('../toolbox')
+from helper_functions import score_function, read_particles, read_snapshot
+
 from mpi4py import MPI
 comm = MPI.COMM_WORLD
 comm_rank = comm.Get_rank()
 comm_size = comm.Get_size()
 
-import h5py
 import numpy as np
-
-import virgo.mpi.util
 import virgo.mpi.parallel_hdf5 as phdf5
 import virgo.mpi.parallel_sort as psort
 
-def read_snapshot(snapshot_file, snap_nr, particle_ids):
+def check_tracing_resolved_subgroups(basedir, hbt_nr, snap_nr, snapshot_file):
     """
-    Read particle properties for the specified particle IDs.
-    Returns a dict of arrays.
-    """
-
-    # Datasets to pass through from the snapshot
-    passthrough_datasets = ("FOFGroupIDs",)
-    
-    # Sub in the snapshot number
-    from virgo.util.partial_formatter import PartialFormatter
-    formatter = PartialFormatter()
-    filenames = formatter.format(snapshot_file, snap_nr=snap_nr, file_nr=None)
-
-    # Determine what particle types we have in the snapshot
-    if comm_rank == 0:
-        ptypes = []
-        with h5py.File(filenames.format(file_nr=0), "r") as infile:
-            nr_types = int(infile["Header"].attrs["NumPartTypes"])
-            nr_parts = infile["Header"].attrs["NumPart_Total"]
-            nr_parts_hw = infile["Header"].attrs["NumPart_Total_HighWord"]
-            for i in range(nr_types):
-                if nr_parts[i] > 0 or nr_parts_hw[i] > 0:
-                    ptypes.append(i)
-    else:
-        ptypes = None
-    ptypes = comm.bcast(ptypes)
-
-    # Read the particle data from the snapshot
-    particle_data = {"Type" : -np.ones(particle_ids.shape, dtype=np.int32)}
-    mf = phdf5.MultiFile(filenames, file_nr_attr=("Header","NumFilesPerSnapshot"), comm=comm)
-    for ptype in ptypes:
-        if ptype == 6:
-            continue # skip neutrinos
-        # Read the IDs of this particle type
-        if comm_rank == 0:
-            print(f"Reading snapshot particle IDs for type {ptype}")
-        snapshot_ids = mf.read(f"PartType{ptype}/ParticleIDs")
-        # For each subhalo particle ID, find matching index in the snapshot (if any)
-        ptr = psort.parallel_match(particle_ids, snapshot_ids, comm=comm)
-        matched = (ptr>=0)
-        # Loop over particle properties to pass through
-        for name in passthrough_datasets:
-            # Read this property from the snapshot
-            snapshot_data = mf.read(f"PartType{ptype}/{name}")
-            # Allocate output array, if we didn't already
-            if name not in particle_data:
-                shape = (len(particle_ids),)+snapshot_data.shape[1:]
-                dtype = snapshot_data.dtype
-                particle_data[name] = -np.ones(shape, dtype=dtype) # initialize to -1 = not found
-            # Look up the value for each subhalo particle
-            # if comm_rank == 0:
-                # print(f"Looking up particle type {ptype} property {name} from snapshot")
-            particle_data[name][matched,...] = psort.fetch_elements(snapshot_data, ptr[matched], comm=comm)
-        
-        # Also store the type of each matched particle
-        particle_data["Type"][matched] = ptype
-
-    # Should have matched all particles, except for merged black holes 
-    # assert np.all(particle_data["Type"] >= 0)
-        
-    return particle_data
-    
-
-def read_particles(filenames, nr_local_subhalos):
-    """
-    Read in the particle IDs belonging to the subhalos on this MPI
-    rank from the specified SubSnap files. Returns a single array with
-    the concatenated IDs from all local subhalos in the order they
-    appear in the SubSnap files.
-    """
-
-    # First determine how many subhalos are in each SubSnap file
-    if comm_rank == 0:
-        subhalos_per_file = []
-        nr_files = 1
-        file_nr = 0
-        while file_nr < nr_files:
-            with h5py.File(filenames.format(file_nr=file_nr), "r") as infile:
-                subhalos_per_file.append(infile["Subhalos"].shape[0])
-                nr_files = int(infile["NumberOfFiles"][...])
-            file_nr += 1
-    else:
-        subhalos_per_file = None
-        nr_files = None
-    nr_files = comm.bcast(nr_files)
-    subhalos_per_file = np.asarray(comm.bcast(subhalos_per_file), dtype=int)
-    first_subhalo_in_file = np.cumsum(subhalos_per_file) - subhalos_per_file
-    
-    # Determine offset to first subhalo this rank reads
-    first_local_subhalo = comm.scan(nr_local_subhalos) - nr_local_subhalos
-
-    # Loop over all files
-    particle_ids = []
-    for file_nr in range(nr_files):
-
-        # Find range of subhalos this rank read from this file
-        i1 = first_local_subhalo - first_subhalo_in_file[file_nr]
-        i2 = i1 + nr_local_subhalos
-        i1 = max(0, i1)
-        i2 = min(subhalos_per_file[file_nr], i2)
-
-        # Read subhalo particle IDs, if there are any in this file for this rank
-        if i2 > i1:
-            with h5py.File(filenames.format(file_nr=file_nr), "r") as infile:
-                particle_ids.append(infile["SubhaloParticles"][i1:i2])
-
-    if len(particle_ids) > 0:
-        # Combine arrays from different files
-        particle_ids = np.concatenate(particle_ids)
-        # Combine arrays from different subhalos
-        particle_ids = np.concatenate(particle_ids)        
-    else:
-        # Some ranks may have read zero files
-        particle_ids = None
-    particle_ids = virgo.mpi.util.replace_none_with_zero_size(particle_ids, comm=comm)
-    
-    return particle_ids
-
-def rank_weight(rank):
-    '''
-    Function used to weigh the contribution of each particle towards a host FOF
-    decision, based on its boundness ranking.
+    This function checks if the internally assigned host FOF of resolved 
+    subhaloes is done correctly. 
 
     Parameters
-    -----------
-    rank : np.ndarray
-        Boundness ranking of the particle, in the previous output..
+    ----------
+    basedir : str    
+        Location of the HBT catalogues.
+    hbt_nr : int
+        Snapshot index to select the ids used for tracing. Their FOF hosts will
+        be used to determine the host of the subgroup we are tracking.
+    snap_nr : int
+        Snapshot number to test. Not equal to hbt_nr if the catalogues have only
+        been created for a subset of snapshots.
+    snapshot_file : str
+        Path to the snapshots in the form SNAPSHOT_BASE_NAME_{snap_nr:04d}.{file_nr}.hdf5
 
     Returns
-    -----------
-    np.ndarray
-        The weight of the particle used to score candidates.
-    '''
-    return 1 / (1 + np.sqrt(rank))
-
-def score_function(fof_groups):
-    '''
-    Score each candidate FOF based on several particles. Return the host with the
-    highest score.
-
-    Parameters
-    -----------
-    fof_groups : np.ndarray
-        The FOF group membership of a series of particles, previously sorted
-        descending boundness ranking order.
-
-    Returns
-    -----------
-    int
-        The highest scoring FOF candidate.
-    '''
-    weights = rank_weight(np.arange(len(fof_groups)))
-    
-    unique_candidates = np.unique(fof_groups)
-    scores = {}
-    for cand in unique_candidates:
-        scores[cand] = np.sum(weights[fof_groups == cand])
-
-    temp = max(scores.values())
-    result = [key for key in scores if scores[key] == temp][0] 
-    if result == 2147483647: result =-1 
-    return result
-
-def sort_hbt_output(basedir, hbt_nr, snap_nr, snapshot_file):
-
+    -------
+    total_number_mistracks: int
+        Number of resolved subgroups whose internally assigned FOF disagrees 
+        with the one we just found.
     """
-    This reorganizes a set of HBT SubSnap files into a single file which
-    contains one HDF5 dataset for each subhalo property. Subhalos are written
-    in order of TrackId.
 
-    Particle IDs in groups can be optionally copied to the output.
-    """
-    
     # Read in the input subhalos
     if comm_rank == 0:
         print(f"Testing HBTplus tracing of resolved subgroups between snapshot index {hbt_nr} and {hbt_nr + 1}")
@@ -279,7 +132,7 @@ def sort_hbt_output(basedir, hbt_nr, snap_nr, snapshot_file):
         print()
         print(f"Reading particle information.")
     
-    particle_data = read_snapshot(snapshot_file, snap_nr + 1, particle_ids)
+    particle_data = read_snapshot(snapshot_file, snap_nr + 1, particle_ids,("FOFGroupIDs",))
     
     if comm_rank == 0:
         print(f"Done reading particle information.")
@@ -333,11 +186,13 @@ def sort_hbt_output(basedir, hbt_nr, snap_nr, snapshot_file):
     local_number_disagreements = np.sum((fofs != fof_decisions[:,1]))
 
     # Check across all ranks across
-    total_number_disagreements = comm.allreduce(local_number_disagreements)
+    total_number_mistracks = comm.allreduce(local_number_disagreements)
     total_number_checks = comm.allreduce(len(fof_decisions))
     
     if comm_rank == 0:
-        print(f"{total_number_disagreements} out of {total_number_checks} disagree.")
+        print(f"{total_number_mistracks} out of {total_number_checks} disagree.")
+
+    return total_number_mistracks
 
 if __name__ == "__main__":
 
@@ -351,4 +206,4 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
 
-    sort_hbt_output(**vars(args))
+    check_tracing_resolved_subgroups(**vars(args))
